+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Fixed
 - Expand `~` when set in `scgi_url`
   (https://github.com/kannibalox/pyrosimple/issues/34)
+- `rtcontrol`: Under some circumstances, `~/.rtorrent` would not be checked for a connection
 
 ## [2.8.0] - 2023-04-29